Implement set budget feature
Related features include parsing of budget and exceptions handling,
creation of budget command for execution of setting the budget,
and the addition of help and UI messages.

+++ src/main/java/seedu/duke/data/Budget.java
@@ -13,14 +13,14 @@ import static seedu.duke.common.Constants.MIN_BUDGET_VALUE;
 public class Budget {
     //@@author wcwy
     // Default value of the monthly budget is $1000
-    private static int budget = 1000;
+    private static long budget = 1000;
 
     /**
      * Retrieves the budget value set for the current month.
      *
      * @return The budget value set by the user.
      */
-    public static int getBudget() {
+    public static long getBudget() {
         return budget;
     }
 
@@ -29,7 +29,7 @@ public class Budget {
      *
      * @param budget The new value for budget.
      */
-    public static void setBudget(int budget) {
+    public static void setBudget(long budget) {
         Budget.budget = budget;
     }
